requirements.txt
cffi==1.14.6
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==2.1.0
Werkzeug==2.0.1
